requirements.txt
fastmcp>=2.0.0
httpx>=0.27.0
python-dotenv>=1.0.0
openai>=1.0.0
fastapi>=0.100.0
uvicorn>=0.20.0
